requirements.txt
python-telegram-bot==20.3
boto3==1.28.57
requests
